requirements.txt
# --- Core LangChain Framework ---
langchain==0.3.7
langchain-core>=0.3.35,<0.4.0
langchain-community==0.3.5

# --- HuggingFace Integration ---
langchain-huggingface
transformers
huggingface-hub

# --- Vector Store (FAISS) ---
faiss-cpu==1.13.1

# --- Embeddings ---
sentence-transformers==4.1.0

# --- Data Handling ---
pypdf==5.6.1

# --- Environment ---
python-dotenv==1.1.0

# --- API Framework ---
fastapi==0.115.0
uvicorn==0.30.1

# --- groq SDK ---
groq
langchain-groq

